+++ tests/TestCase/Command/IndexCommandTest.php
@@ -14,44 +14,33 @@ declare(strict_types=1);
  */
 namespace DatabaseBackup\Test\TestCase\Command;
 
-use DatabaseBackup\TestSuite\TestCase;
-use MeTools\TestSuite\ConsoleIntegrationTestTrait;
+use DatabaseBackup\TestSuite\CommandTestCase;
 
 /**
  * IndexCommandTest class
  */
-class IndexCommandTest extends TestCase
+class IndexCommandTest extends CommandTestCase
 {
-    use ConsoleIntegrationTestTrait;
-
     /**
-     * @var string
-     */
-    protected $command = 'database_backup.index -v';
-
-    /**
-     * Test for `execute()` method
      * @test
+     * @uses \DatabaseBackup\Command\IndexCommand::execute()
      */
     public function testExecute(): void
     {
-        $this->createSomeBackups();
-        $this->exec($this->command);
-        $this->assertExitWithSuccess();
+        $command = 'database_backup.index -v';
+
+        $files = createSomeBackups();
+        $this->exec($command);
+        $this->assertExitSuccess();
         $this->assertOutputContains('Backup files found: 3');
         $this->assertOutputRegExp('/backup\.sql\.gz\s+|\s+sql\.gz\s+|\s+gzip\s+|\s+[\d\.]+ \w+\s+|\s+[\d\/]+, [\d:]+ (AP)M/');
         $this->assertOutputRegExp('/backup\.sql\.bz2\s+|\s+sql\.bz2\s+|\s+bzip2\s+|\s+[\d\.]+ \w+\s+|\s+[\d\/]+, [\d:]+ (AP)M/');
         $this->assertOutputRegExp('/backup\.sq\s+|\s+sql\s+|\s+|\s+[\d\.]+ \w+\s+|\s+[\d\/]+, [\d:]+ (AP)M/');
-    }
 
-    /**
-     * Test for `execute()` method, with no backups
-     * @test
-     */
-    public function testExecuteNoBackups(): void
-    {
-        $this->exec($this->command);
-        $this->assertExitWithSuccess();
+        //With no backups
+        array_map('unlink', $files);
+        $this->exec($command);
+        $this->assertExitSuccess();
         $this->assertOutputContains('Connection: test');
         $this->assertOutputRegExp('/Driver: (Mysql|Postgres|Sqlite)/');
         $this->assertOutputContains('Backup files found: 0');